benchmarks: move all benchmarks to their own file
By having everything in one file it reduces clutter in the codebase,
will make it easier to unify benchmark-generation code, and (IMO most
importantly) lets you write `cargo bench benchmarks` and then it'll only
run the benchmarks instead of outputting a gazillion "ignored" lines
telling you that it's not benchmarking random shit that aren't
benchmarks.

Don't re-format stuff so it'll be obvious in the diff that this is a
pure code move. Also we are going to rewrite the benchmarks in the next
commit anyway.
