+++ src/lib.rs
@@ -113,6 +113,8 @@ mod macros;
 #[macro_use]
 mod pub_macros;
 
+#[cfg(bench)]
+mod benchmarks;
 mod blanket_traits;
 pub mod descriptor;
 pub mod expression;
