+++ src/policy/compiler.rs
@@ -1649,52 +1649,3 @@ mod tests {
         }
     }
 }
-
-#[cfg(bench)]
-mod benches {
-    use std::str::FromStr;
-
-    use test::{black_box, Bencher};
-
-    use super::{CompilerError, Concrete};
-    use crate::miniscript::Tap;
-    use crate::prelude::*;
-    use crate::Miniscript;
-    type TapMsRes = Result<Miniscript<String, Tap>, CompilerError>;
-    #[bench]
-    pub fn compile_basic(bh: &mut Bencher) {
-        let h = (0..64).map(|_| "a").collect::<String>();
-        let pol = Concrete::<String>::from_str(&format!(
-            "and(thresh(2,and(sha256({}),or(sha256({}),pk(A))),pk(B),pk(C),pk(D),sha256({})),pk(E))",
-            h, h, h
-        ))
-        .expect("parsing");
-        bh.iter(|| {
-            let pt: TapMsRes = pol.compile();
-            black_box(pt).unwrap();
-        });
-    }
-
-    #[bench]
-    pub fn compile_large(bh: &mut Bencher) {
-        let h = (0..64).map(|_| "a").collect::<String>();
-        let pol = Concrete::<String>::from_str(
-            &format!("or(pk(L),thresh(9,sha256({}),pk(A),pk(B),and(or(pk(C),pk(D)),pk(E)),after(100),pk(F),pk(G),pk(H),pk(I),and(pk(J),pk(K))))", h)
-        ).expect("parsing");
-        bh.iter(|| {
-            let pt: TapMsRes = pol.compile();
-            black_box(pt).unwrap();
-        });
-    }
-
-    #[bench]
-    pub fn compile_xlarge(bh: &mut Bencher) {
-        let pol = Concrete::<String>::from_str(
-            "or(pk(A),thresh(4,pk(B),older(100),pk(C),and(after(100),or(pk(D),or(pk(E),and(pk(F),thresh(2,pk(G),or(pk(H),and(thresh(5,pk(I),or(pk(J),pk(K)),pk(L),pk(M),pk(N),pk(O),pk(P),pk(Q),pk(R),pk(S),pk(T)),pk(U))),pk(V),or(and(pk(W),pk(X)),pk(Y)),after(100)))))),pk(Z)))"
-        ).expect("parsing");
-        bh.iter(|| {
-            let pt: TapMsRes = pol.compile();
-            black_box(pt).unwrap();
-        });
-    }
-}
